Replace copper n and k with new improved table
The new values are based on the Cu n&k determination of Babar
and Weaver with modified HO parameters to match local thin
film data.

Two preset HO are given for Copper, the first correspond to
the original Babar-Weaver determination while the second
corresponds to the modified version included in the library.