Implement clipboard operation to copy plot as image
The current implementation works on windows but probably not
on linux because it assumes the host system use BMP image
format.

The image is not actually stored when the application gain the
clipboard but is generated on the fly when the clipboard
request is received. Probably is more correct to store the
image when the clipboard in gained.

Based on hints found on FOX Mailing list:

http://fox-toolkit.2306850.n4.nabble.com/Clipboard-Image-Data-td4006820.html