Implement support to fit the bandwidth
Calculate derivative with bandwidth using the derivatives with the
wavelengths. These latters are computed using the derivatives with the
thicknesses.

All the computations are done in the approximation of constant RI
versus lambda for the films. In reality the RI depends on lambda but the
approx is ok in general because the lambda integration is done in a small
interval.