Do not use VLA but use instead a specific class instead
Required to be compliant to C++14 and to compile with clang++ since VLA
is not supported.

It does use the class StackArray that use stack only for small size and
revert to heap allocated memory for bigger size.

The new approach is more safe because stack cannot be smashed due to
big array size.

Main disadvantage is that require a call to .data() method to obtain the
pointer.