Ensure that a minimum number of points in ensured for fit
When minimizing the number of points in the sectrum ensure that a minimum
number of points in provided. Iterate a few times with more accuracy to
increase the number of points and do no sampling if goal cannot be reached.

Needed to prevent GSL library error with non-linear fit if number of
experimental points is less than te number of fit parameters.