Improve Tauc-Lorentz model to work when C > 2 * E0 (standard form)
Context
-------

In the previous version of Tauc-Lorentz, the one written by hand, a
special form of epsilon1 was given for the case when C > 2 * E0 to avoid
NaNs. The problem was that when C > 2 * E0 the derived value

alpha = sqrt(4*E0^2 - C^2)

becomes imaginary.

When the Tauc-Lorentz was ported to sympy this form for C > 2 * E0 was
not implemented and, as a result, the Tauc-Lorentz was giving NaNs for
C > 2 * E0.

Changes
=======

Implement with the sympy generator the special form for the case when
C > 2 * E0. In sympy it easily done by writing the two different forms
and then join them in a simple expression using the Piecewise function.

In addition, the expression to evaluate n for Tauc-Lorentz without
derivatives was using sympy but it was hard-coded and the script was not
given. Now the sympy script to generate the Tauc-Lorentz without
derivative is given.