Implement test for 2nd order HO model
Provides a HO-like model that corresponds to an HO model with
phi = 0 but it is Kramers-Kronig correct and its imaginary
part is odd with energy (or frequency).

The imaginary part will always be partly positive and partly
negative. This is a feature of the model. It is supposed to
participate to the epsilon (dielectric constant) of a
material in addition to other terms so that the resulting
epsilon have an always positive imaginary part.